create useLogin hook
