+++ frontend/src/screens/Welcome/Welcome.tsx
@@ -1,107 +1,21 @@
-import { useEffect, useContext } from "react";
-import { useNavigate } from "react-router-dom";
-import { User, UserContext } from "../../contexts/userContext";
 import { useAllUsersQuery } from "../../graphql/allUsers.graphql.types";
-import { useCurrentUserLazyQuery } from "../../graphql/currentUser.graphql.types";
 import { useUser } from "../../hooks/common/useUser";
-import { pathsGenerator } from "../../router/paths";
 import { Styles } from "../../utils/Styles";
-import { Roles, defaultUnauthenticatedUser } from "../../utils/types";
-import { signInWithEmailAndPassword } from "firebase/auth";
-import { auth } from "../../../firebaseConfig";
-import { UsersRolesType } from "../../__generated__/schema.graphql.types";
-import Cookies from "js-cookie";
+import { Roles } from "../../utils/types";
 import { LoginForm } from "../../components/Welcome/LoginForm";
 import { SelectedUserInfo } from "../../components/Welcome/SelectedUserInfo";
 import { UsersListWithFilter } from "../../components/Welcome/UsersListWithFilter/UsersListWithFilter";
+import { useLogin } from "../../hooks/auth/useLogin";
 
 export const Welcome = () => {
-  const navigate = useNavigate();
-  const context = useContext(UserContext);
-
-  const { user: selectedUser, setUser } = useUser();
+  const { user: selectedUser } = useUser();
   const { loading, error, data } = useAllUsersQuery({
     context: {
       headers: { "x-hasura-role": Roles.UNAUTHENTICATED_USER },
     },
   });
 
-  const [fetchCurrentUser, { data: currentUserData }] =
-    useCurrentUserLazyQuery();
-
-  useEffect(() => {
-    if (currentUserData) {
-      let role = Roles.UNAUTHENTICATED_USER;
-      switch (currentUserData.getCurrentUser?.role) {
-        case UsersRolesType.Student:
-          role = Roles.STUDENT;
-          break;
-        case UsersRolesType.Teacher:
-          role = Roles.TEACHER;
-          break;
-        case UsersRolesType.Coordinator:
-          role = Roles.COORDINATOR;
-          break;
-      }
-      const userData = {
-        userId:
-          currentUserData.getCurrentUser?.userId ||
-          defaultUnauthenticatedUser.userId,
-        role: role || defaultUnauthenticatedUser.role,
-        nick:
-          currentUserData.getCurrentUser?.nick ||
-          defaultUnauthenticatedUser.nick,
-      };
-      setUser(userData);
-      // Store user data in cookies
-      Cookies.set("user", JSON.stringify(userData));
-      navigate(
-        role === Roles.STUDENT
-          ? pathsGenerator.student.StudentProfile
-          : pathsGenerator.teacher.Groups,
-      );
-    }
-  }, [currentUserData, setUser, navigate]);
-
-  const handleUserSelect = (user: User) => {
-    if (!context) {
-      throw new Error("useContext must be used within a UserProvider");
-    }
-    const userData = {
-      userId: user?.userId || defaultUnauthenticatedUser.userId,
-      role: user?.role || defaultUnauthenticatedUser.role,
-      nick: user?.nick || defaultUnauthenticatedUser.nick,
-    };
-    setUser(user);
-    const token = "Bypass" + user.userId;
-    Cookies.set("token", token || "");
-    Cookies.set("user", JSON.stringify(userData));
-    navigate(
-      user.role === Roles.STUDENT
-        ? pathsGenerator.student.StudentProfile
-        : pathsGenerator.teacher.Groups,
-    );
-  };
-
-  const handleLogin = async (email: string, password: string) => {
-    if (!context) {
-      throw new Error("useContext must be used within a UserProvider");
-    }
-
-    try {
-      let token: string | undefined;
-      if (password.length < 4) {
-        token = "Bypass" + password;
-      } else {
-        await signInWithEmailAndPassword(auth, email, password);
-        token = await auth.currentUser?.getIdToken();
-      }
-      Cookies.set("token", token || "");
-      fetchCurrentUser();
-    } catch (error) {
-      throw new Error("Invalid email or password.");
-    }
-  };
+  const { handleLogin, handleUserSelect } = useLogin();
 
   if (loading) return <div>Ładowanie...</div>;
   if (error) return <div>Błąd podczas ładowania użytkowników.</div>;
